Small refactor to use networking packet handler
