+++ src/main/java/com/fouristhenumber/utilitiesinexcess/CommonProxy.java
@@ -1,7 +1,7 @@
 package com.fouristhenumber.utilitiesinexcess;
 
-import com.fouristhenumber.utilitiesinexcess.utils.SoundVolumeChecks;
 import com.fouristhenumber.utilitiesinexcess.network.PacketHandler;
+import com.fouristhenumber.utilitiesinexcess.utils.SoundVolumeChecks;
 
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
