+++ src/main/java/com/fouristhenumber/utilitiesinexcess/UtilitiesInExcess.java
@@ -10,7 +10,6 @@ import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityDrum;
 import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityRainMuffler;
 import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityRedstoneClock;
 import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntitySoundMuffler;
-import com.fouristhenumber.utilitiesinexcess.network.PacketRainMuffledSync;
 import com.fouristhenumber.utilitiesinexcess.utils.EventHandler;
 
 import cpw.mods.fml.common.Mod;
@@ -19,9 +18,7 @@ import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.event.FMLServerStartingEvent;
-import cpw.mods.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import cpw.mods.fml.common.registry.GameRegistry;
-import cpw.mods.fml.relauncher.Side;
 
 @Mod(
     modid = UtilitiesInExcess.MODID,
@@ -34,8 +31,6 @@ public class UtilitiesInExcess {
     public static final String MODID = "utilitiesinexcess";
     public static final Logger LOG = LogManager.getLogger(MODID);
 
-    public static SimpleNetworkWrapper networkWrapper;
-
     @SidedProxy(
         clientSide = "com.fouristhenumber.utilitiesinexcess.ClientProxy",
         serverSide = "com.fouristhenumber.utilitiesinexcess.CommonProxy")
@@ -44,9 +39,6 @@ public class UtilitiesInExcess {
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
         proxy.preInit(event);
-
-        networkWrapper = new SimpleNetworkWrapper(MODID);
-        networkWrapper.registerMessage(PacketRainMuffledSync.class, PacketRainMuffledSync.class, 1, Side.CLIENT);
     }
 
     @Mod.EventHandler
