+++ src/main/java/com/fouristhenumber/utilitiesinexcess/network/PacketHandler.java
@@ -1,6 +1,7 @@
 package com.fouristhenumber.utilitiesinexcess.network;
 
 import com.fouristhenumber.utilitiesinexcess.UtilitiesInExcess;
+import com.fouristhenumber.utilitiesinexcess.network.client.PacketRainMuffledSync;
 import com.fouristhenumber.utilitiesinexcess.network.client.ParticlePacket;
 
 import cpw.mods.fml.common.network.simpleimpl.SimpleNetworkWrapper;
@@ -13,5 +14,7 @@ public class PacketHandler {
     public static void init() {
         int packetId = 0;
         INSTANCE.registerMessage(ParticlePacket.Handler.class, ParticlePacket.class, packetId++, Side.CLIENT);
+        INSTANCE
+            .registerMessage(PacketRainMuffledSync.Handler.class, PacketRainMuffledSync.class, packetId++, Side.CLIENT);
     }
 }
